input: move string parser into separate module
