+++ pym/bob/input.py
@@ -18,10 +18,10 @@ from . import BOB_VERSION, BOB_INPUT_HASH, DEBUG
 from .errors import BuildError, ParseError
 from .scm import CvsScm, GitScm, SvnScm, UrlScm, ScmOverride, auditFromDir
 from .state import BobState
+from .stringparser import checkGlobList, Env, DEFAULT_STRING_FUNS
 from .tty import colorize, WarnOnce
 from .utils import asHexStr, joinScripts, sliceString, compareVersion, binStat, updateDicRecursive
 from abc import ABCMeta, abstractmethod
-from collections.abc import MutableMapping
 from base64 import b64encode
 from itertools import chain
 from glob import glob
@@ -70,292 +70,6 @@ def maybeGlob(pattern):
         return [ __maybeGlob(p) for p in pattern ]
     else:
         return None
-
-def checkGlobList(name, allowed):
-    if allowed is None: return True
-    ok = False
-    for pred in allowed: ok = pred(ok, name)
-    return ok
-
-def _isFalse(val):
-    return val.strip().lower() in [ "", "0", "false" ]
-
-class StringParser:
-    """Utility class for complex string parsing/manipulation"""
-
-    def __init__(self, env, funs, funArgs):
-        self.env = env
-        self.funs = funs
-        self.funArgs = funArgs
-
-    def parse(self, text):
-        """Parse the text and make substitutions"""
-        if all((c not in text) for c in '\\\"\'$'):
-            return text
-        else:
-            self.text = text
-            self.index = 0
-            self.end = len(text)
-            return self.getString()
-
-    def nextChar(self):
-        """Get next character"""
-        i = self.index
-        if i >= self.end:
-            raise ParseError('Unexpected end of string')
-        self.index += 1
-        return self.text[i:i+1]
-
-    def nextToken(self, extra=None):
-        delim=['\"', '\'', '$']
-        if extra: delim.extend(extra)
-
-        # EOS?
-        i = start = self.index
-        if i >= self.end:
-            return None
-
-        # directly on delimiter?
-        if self.text[i] in delim:
-            self.index = i+1
-            return self.text[i]
-
-        # scan
-        tok = []
-        while i < self.end:
-            if self.text[i] in delim: break
-            if self.text[i] == '\\':
-                tok.append(self.text[start:i])
-                start = i = i + 1
-                if i >= self.end:
-                    raise ParseError("Unexpected end after escape")
-            i += 1
-        tok.append(self.text[start:i])
-        self.index = i
-        return "".join(tok)
-
-    def getSingleQuoted(self):
-        i = self.index
-        while i < self.end:
-            if self.text[i] == "'":
-                i += 1
-                break
-            i += 1
-        if i >= self.end:
-            raise ParseError("Missing closing \"'\"")
-        ret = self.text[self.index:i-1]
-        self.index = i
-        return ret
-
-    def getString(self, delim=[None], keep=False):
-        s = []
-        tok = self.nextToken(delim)
-        while tok not in delim:
-            if tok == '"':
-                s.append(self.getString(['"']))
-            elif tok == '\'':
-                s.append(self.getSingleQuoted())
-            elif tok == '$':
-                tok = self.nextChar()
-                if tok == '{':
-                    s.append(self.getVariable())
-                elif tok == '(':
-                    s.append(self.getCommand())
-                else:
-                    raise ParseError("Invalid $-subsitituion")
-            elif tok == None:
-                if None not in delim:
-                    raise ParseError('Unexpected end of string')
-                break
-            else:
-                s.append(tok)
-            tok = self.nextToken(delim)
-        else:
-            if keep: self.index -= 1
-        return "".join(s)
-
-    def getVariable(self):
-        # get variable name
-        varName = self.getString([':', '-', '+', '}'], True)
-
-        # process?
-        op = self.nextChar()
-        unset = varName not in self.env
-        if op == ':':
-            # or null...
-            if not unset: unset = self.env[varName] == ""
-            op = self.nextChar()
-
-        if op == '-':
-            default = self.getString(['}'])
-            if unset:
-                return default
-            else:
-                return self.env[varName]
-        elif op == '+':
-            alternate = self.getString(['}'])
-            if unset:
-                return ""
-            else:
-                return alternate
-        elif op == '}':
-            if varName not in self.env:
-                raise ParseError("Unset variable: " + varName)
-            return self.env[varName]
-        else:
-            raise ParseError("Unterminated variable: " + str(op))
-
-    def getCommand(self):
-        words = []
-        delim = [",", ")"]
-        while True:
-            word = self.getString(delim, True)
-            words.append(word)
-            end = self.nextChar()
-            if end == ")": break
-
-        if len(words) < 1:
-            raise ParseError("Expected function name")
-        cmd = words[0]
-        del words[0]
-
-        if cmd not in self.funs:
-            raise ParseError("Unknown function: "+cmd)
-
-        return self.funs[cmd](words, env=self.env, **self.funArgs)
-
-class Env(MutableMapping):
-    def __init__(self, other={}):
-        self.data = dict(other)
-        self.funs = []
-        self.funArgs = {}
-        self.touched = [ set() ]
-
-    def __touch(self, key):
-        for i in self.touched: i.add(key)
-
-    def __contains__(self, key):
-        self.__touch(key)
-        return key in self.data
-
-    def __delitem__(self, key):
-        del self.data[key]
-
-    def __eq__(self, other):
-        if isinstance(other, Env):
-            return self.data == other.data
-        else:
-            return self.data == other
-
-    def __getitem__(self, key):
-        self.__touch(key)
-        return self.data[key]
-
-    def __iter__(self):
-        raise NotImplementedError("iter() not supported")
-
-    def __len__(self):
-        return len(self.data)
-
-    def __ne__(self, other):
-        if isinstance(other, Env):
-            return self.data != other.data
-        else:
-            return self.data != other
-
-    def __setitem__(self, key, value):
-        self.data[key] = value
-
-    def clear(self):
-        self.data.clear()
-
-    def copy(self):
-        ret = Env(self.data)
-        ret.funs = self.funs
-        ret.funArgs = self.funArgs
-        ret.touched = self.touched
-        return ret
-
-    def get(self, key, default=None):
-        self.__touch(key)
-        return self.data.get(key, default)
-
-    def items(self):
-        raise NotImplementedError("items() not supported")
-
-    def keys(self):
-        raise NotImplementedError("keys() not supported")
-
-    def pop(self, key, default=None):
-        raise NotImplementedError("pop() not supported")
-
-    def popitem(self):
-        raise NotImplementedError("popitem() not supported")
-
-    def update(self, other):
-        self.data.update(other)
-
-    def values(self):
-        raise NotImplementedError("values() not supported")
-
-    def derive(self, overrides = {}):
-        ret = self.copy()
-        ret.data.update(overrides)
-        return ret
-
-    def detach(self):
-        return self.data.copy()
-
-    def setFuns(self, funs):
-        self.funs = funs
-
-    def setFunArgs(self, funArgs):
-        self.funArgs = funArgs
-
-    def prune(self, allowed):
-        if allowed is None:
-            return self.copy()
-        else:
-            ret = Env()
-            ret.data = { key : self.data[key] for key in (set(self.data.keys()) & allowed) }
-            ret.funs = self.funs
-            ret.funArgs = self.funArgs
-            ret.touched = self.touched
-            return ret
-
-    def filter(self, allowed):
-        if allowed is None:
-            return self.copy()
-        else:
-            ret = Env()
-            ret.data = { key : value for (key, value) in self.data.items()
-                if checkGlobList(key, allowed) }
-            ret.funs = self.funs
-            ret.funArgs = self.funArgs
-            ret.touched = self.touched
-            return ret
-
-    def substitute(self, value, prop):
-        try:
-            return StringParser(self, self.funs, self.funArgs).parse(value)
-        except ParseError as e:
-            raise ParseError("Error substituting {}: {}".format(prop, str(e.slogan)))
-
-    def evaluate(self, condition, prop):
-        if condition is None:
-            return True
-
-        s = self.substitute(condition, "condition on "+prop)
-        return not _isFalse(s)
-
-    def touchReset(self):
-        self.touched = self.touched + [ set() ]
-
-    def touch(self, keys):
-        for k in keys: self.__touch(k)
-
-    def touchedKeys(self):
-        return self.touched[-1]
 
 
 class PluginProperty:
@@ -2113,72 +1827,6 @@ class PackageMatcher:
         inputTools.touch(self.tools.keys())
 
 
-def funEqual(args, **options):
-    if len(args) != 2: raise ParseError("eq expects two arguments")
-    return "true" if (args[0] == args[1]) else "false"
-
-def funNotEqual(args, **options):
-    if len(args) != 2: raise ParseError("ne expects two arguments")
-    return "true" if (args[0] != args[1]) else "false"
-
-def funNot(args, **options):
-    if len(args) != 1: raise ParseError("not expects one argument")
-    return "true" if _isFalse(args[0]) else "false"
-
-def funOr(args, **options):
-    for arg in args:
-        if not _isFalse(arg):
-            return "true"
-    return "false"
-
-def funAnd(args, **options):
-    for arg in args:
-        if _isFalse(arg):
-            return "false"
-    return "true"
-
-def funMatch(args, **options):
-    try:
-        [2, 3].index(len(args))
-    except ValueError:
-        raise ParseError("match expects either two or three arguments")
-
-    flags = 0
-    if len(args) == 3:
-        if args[2] == 'i':
-            flags = re.IGNORECASE
-        else:
-            raise ParseError('match only supports the ignore case flag "i"')
-
-    if re.search(args[1],args[0],flags):
-        return "true"
-    else:
-        return "false"
-
-def funIfThenElse(args, **options):
-    if len(args) != 3: raise ParseError("if-then-else expects three arguments")
-    if _isFalse(args[0]):
-        return args[2]
-    else:
-        return args[1]
-
-def funSubst(args, **options):
-    if len(args) != 3: raise ParseError("subst expects three arguments")
-    return args[2].replace(args[0], args[1])
-
-def funStrip(args, **options):
-    if len(args) != 1: raise ParseError("strip expects one argument")
-    return args[0].strip()
-
-def funSandboxEnabled(args, sandbox, **options):
-    if len(args) != 0: raise ParseError("is-sandbox-enabled expects no arguments")
-    return "true" if ((sandbox is not None) and sandbox.isEnabled()) else "false"
-
-def funToolDefined(args, tools, **options):
-    if len(args) != 1: raise ParseError("is-tool-defined expects one argument")
-    return "true" if (args[0] in tools) else "false"
-
-
 class ArchiveValidator:
     def __init__(self):
         self.__validTypes = schema.Schema({'backend': schema.Or('none', 'file', 'http', 'shell')},
@@ -2312,19 +1960,7 @@ class RecipeSet:
         self.__properties = {}
         self.__states = {}
         self.__cache = YamlCache()
-        self.__stringFunctions = {
-            "eq" : funEqual,
-            "or" : funOr,
-            "and" : funAnd,
-            "if-then-else" : funIfThenElse,
-            "is-sandbox-enabled" : funSandboxEnabled,
-            "is-tool-defined" : funToolDefined,
-            "ne" : funNotEqual,
-            "not" : funNot,
-            "strip" : funStrip,
-            "subst" : funSubst,
-            "match" : funMatch,
-        }
+        self.__stringFunctions = DEFAULT_STRING_FUNS.copy()
         self.__plugins = {}
         self.__commandConfig = {}
         self.__policies = {
